Created a Manhwa calss in the run_utilities, but the chapter that was posted today isn't showing up. Diagnose later
